backend/requirements.txt
fastapi==0.109.0
pydantic==2.5.3
pymongo==4.6.1
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
python-dotenv==1.0.0
uvicorn==0.25.0
email-validator==2.1.0
jinja2==3.1.3
